Introduction

Most of the functions work on a SymbolicAWEModel object. For this, the variable s is used. Such a variable can be created with the lines:

set = Settings("system_ram.yaml")
s = SymbolicAWEModel(set)

Functions with an "!" as last character of the function name modify one of more of their parameters, in this context mostly the variable s.

Input functions

SymbolicAWEModels.set_depower_steering! — Function

set_depower_steering!(s::SymbolicAWEModel, depower, steering) -> Nothing

Set kite depower and steering by adjusting tether lengths. Depower controls angle of attack, steering controls left/right differential. Values are scaled by minimum chord length.

SymbolicAWEModels.set_v_wind_ground! — Function

set_v_wind_ground!(s::SymbolicAWEModel, v_wind_gnd=s.set.v_wind, upwind_dir=-π/2) -> Nothing

Set ground wind speed (m/s) and upwind direction (radians). Direction: 0=north, π/2=east, π=zouth, -π/2=west (default).

Output functions

SymbolicAWEModels.unstretched_length — Function

Returns the unstretched tether length of the symbolic AWE model.

SymbolicAWEModels.tether_length — Function

Returns the current tether length of the symbolic AWE model.

SymbolicAWEModels.calc_height — Function

Returns the height (z-position) of the wing in the symbolic AWE model.

SymbolicAWEModels.winch_force — Function

Returns the winch force in the symbolic AWE model.

SymbolicAWEModels.spring_forces — Function

Returns the spring forces in the symbolic AWE model.

SymbolicAWEModels.calc_aoa — Function

Calculate and return the angle of attack in rad

SymbolicAWEModels.pos — Function

Returns the position vector of the wing in the symbolic AWE model.

Simulation interface

SymbolicAWEModels.init_sim! — Function

init_sim!(s::SymbolicAWEModel; solver=nothing, adaptive=true, prn=true, 
          precompile=false, remake=false, reload=false, 
          lin_outputs=Num[]) -> OrdinaryDiffEqCore.ODEIntegrator

Initialize a kite power system model.

If a serialized model exists for the current configuration, it will load that model and only update the state variables. Otherwise, it will create a new model from scratch.

Fast path (serialized model exists):

Loads existing ODEProblem from disk

Calls reinit! to update state variables

Sets up integrator with initial settings

Slow path (no serialized model):

Creates symbolic MTK system with all equations

Simplifies system equations

Creates ODEProblem and serializes to disk

Proceeds with fast path

Arguments

s::SymbolicAWEModel: The kite system state object

Keyword arguments

solver: Solver algorithm to use. If nothing, defaults to FBDF() or QNDF() based on s.set.solver.

adaptive::Bool=true: Whether to use adaptive time stepping.

prn::Bool=true: Whether to print progress information.

precompile::Bool=false: Whether to build problem for precompilation.

remake::Bool=false: If true, forces the system to be rebuilt, even if a serialized model exists.

reload::Bool=false: If true, forces the system to reload the serialized model from disk.

lin_outputs::Vector{Num}=Num[]: List of symbolic variables for which to generate a linearization function.

Returns

integrator::OrdinaryDiffEqCore.ODEIntegrator: The initialized ODE integrator.

SymbolicAWEModels.next_step! — Function

next_step!(s::SymbolicAWEModel, integrator::ODEIntegrator; set_values=nothing, upwind_dir=nothing, dt=1/s.set.sample_freq, vsm_interval=1)

Take a simulation step, using the internal integrator.

This function performs the following steps:

Optionally update the set values (control inputs)

Optionally update the upwind direction

Optionally linearize the VSM (Vortex Step Method) model

Step the ODE integrator forward by dt seconds

Check for a successful return code from the integrator

Increment the iteration counter

Arguments

s::SymbolicAWEModel: The kite power system state object

integrator::ODEIntegrator: The ODE integrator to use

Keyword Arguments

set_values=nothing: New values for the set variables (control inputs). If nothing, the current values are used.

dt=1/s.set.sample_freq: Time step size in seconds. Defaults to the inverse of the sample frequency.

vsm_interval=1: Interval (in number of steps) at which to linearize the VSM model. If 0, the VSM model is not linearized.

Returns

Nothing

Helper functions

SymbolicAWEModels.copy_examples — Function

copy_examples()

Copy all example scripts to the folder "examples" (it will be created if it doesn't exist).

SymbolicAWEModels.copy_bin — Function

copy_bin()

Copy the scripts createsysimage and run_julia to the folder "bin" (it will be created if it doesn't exist).