Exported Types

Basic types

SymbolicAWEModels.SimFloat — Type

const SimFloat = Float64

This type is used for all real variables, used in the Simulation. Possible alternatives: Float32, Double64, Dual Other types than Float64 or Float32 do require support of Julia types by the solver.

SymbolicAWEModels.KVec3 — Type

const KVec3 = MVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, mutable.

SymbolicAWEModels.SVec3 — Type

const SVec3 = SVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, immutable.

Struct SymbolicAWEModel

SymbolicAWEModels.SymbolicAWEModel — Type

mutable struct SymbolicAWEModel{S, V, P} <: AbstractKiteModel

State of the kite power system, using a quaternion kite model and three steering lines to the ground. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

V: Vector type, e.g. KVec3

P: number of tether points of the system, 3segments+3

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::KiteUtils.Settings: Reference to the settings struct

sys_struct::SystemStructure: Reference to the point mass system with points, segments, pulleys and tethers

serialized_model::SymbolicAWEModels.SerializedModel

integrator::Union{Nothing, OrdinaryDiffEqCore.ODEIntegrator}: Default: nothing

t_0::Float64: relative start time of the current time interval Default: 0.0

iter::Int64: Number of solve! calls Default: 0

t_vsm::Float64: Default: zero(SimFloat)

t_step::Float64: Default: zero(SimFloat)

set_tether_length::Vector{Float64}: Default: zeros(SimFloat, 3)

It is not advised to access the SymbolicAWEModel fields directly, use the input and output functions.